third_party/blink/web_tests/fast/css/counters/counter-increment-inherit-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.counterValue(document.getElementById('wrapper90')) is '5'
PASS internals.counterValue(document.getElementById('test90')) is '10'
PASS internals.counterValue(document.getElementById('test90a')) is '15'
PASS internals.counterValue(document.getElementById('wrapper91')) is '10'
PASS internals.counterValue(document.getElementById('test91')) is '20'
PASS internals.counterValue(document.getElementById('test91a')) is '30'
